+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - [5124](https://github.com/vegaprotocol/vega/issues/5124) - Add proto serialization for update market proposal
 - [5124](https://github.com/vegaprotocol/vega/issues/5124) - Ensure update market proposal compute a proper auction duration
 - [5172](https://github.com/vegaprotocol/vega/issues/5172) - Add replay protection for validator commands
+- [5181](https://github.com/vegaprotocol/vega/issues/5181) - Ensure Oracle specs handle numbers using `num.Decimal` and `num.Int`
 
 ## 0.49.8